Introduce support for `RelativeTopicPath`
Previously, we added a `RelativeTopicPath` attribute to `TopicListAttribute`, and accounted for it the `TopicListAttributeTopicViewModel`, but we never integrated that functionality into `TopicListViewComponent`. Oops.

As part of that, I rearranged the logic so that the code for the `GetTopics()` call is centralized, instead of handling it separately for `RelativeTopicBase` and `RootTopic`.