Ensure `TopicReferenceAttributeViewComponent` honors current key
The `TopicReferenceAttributeViewComponent` previously had a hard-coded value of `TopicID`. While it is indeed indended primarily to support derived topics (i.e., `TopicID`), it is possible that other attributes may use it in the future if they have similar requirements (i.e., a topic reference that triggers a postback). As such, this maintains the existing functionality while potentially providing more flexibility in the future, if needed.