Expose new `TranslateTopicPointers` option to interface
The `ExportOptions` was expanded to allow the `TranslateTopicPointers` functionality to be optionally disabled as part of the **OnTopic Data Transfer 1.2.0** library, which was integrated as part of the support for working with a clean database (18f676b). As part of that, I missed exposing this option to the `Export()` view in the editor. Whoops!

(In practice, I'm not sure most customers will need this. But it's nice to have for particular edge scenarios.)