Switched to browser's OOTB spell checker
By default, CKEditor disable the browser's OOTB spell checker, in favor of third-party plugins—such as the Spell Check As You Type (SCAYT) plugin we were using. Unfortunately, SCAYT isn't enabled by default, and must be explicitly enabled per session.

SCAYT _could_ be auto-enabled using the `scayt_autoStartup` option. Given that it sends text to a third-party and is advertising supported without a license, however, that's not preferred—and especailly given that the OOTB spell checker is sufficient for most of our customer's needs.