Consolidated formatting onto a single line
This makes the second line longer than the first, which is a bit awkward. But once we fix the `StylesSet.js` configuration, we should be able to get rid of the (largely redundant) `Format` control in favor of simply using `Styles`. At that point, they will be better balanced again.

(Currently, `StylesSet` isn't working properly, so no items are being displayed in the dropdown list.)