Disabled SCAYT plugin
In a previous update (a4790fe), I removed the out-of-the-box Spell Check As You Type (SCAYT) functionality from the toolbar, and reinstated the browser's default spell check feature. I hadn't, however, disabled SCAYT, and thus it was still getting loaded by the browser, even though it wasn't being used. This update actually removes the plugin from the configuration, so it doesn't get loaded.