Removed `toolbar_ontopic` configuration
It seems we had previously configured a version of the toolbar specifically for OnTopic. That might have been useful, but it's not currently being used, and we've been instead applying updates directly to the main `toolbar` configuration. Since we don't have a need to dynamically swap different toolbars for OnTopic currently, I'm removing this. We can always reinstate it in the future if there's a need.