Removed style set customizations
Styles are custom to each customer/implementation. While we could try to establish some common conventions—and, in fact, have in the past—this is likely to be insufficient to a customer's needs, while potentially encouraging the use of markup and classes that aren't actually supported by the target website.

As these are easy enough to overwrite as part of a customer-specific configuration (e.g., by defining their own override of `/Views/Shared/Components/Html/Default.cshtml`) we'll leave that as an implementation option, instead of defining it centrally.