Remove font names
We don't display a font dropdown box, as customer's don't usually want editors redefining the fonts themselves; instead, fonts should be defined by stylesheets. Given that, there's no need to define a custom list of fonts. If we want to reintroduce fonts in the future, we can add this back in—but, likely, if that were ever to happen, it would be on a customer-by-customer basis, with fonts specific to their stylesheet, and not something we'd want to define as part of the default out-of-the-box template.