Modified `InheritedAttributes` to only display inherited attributes
Previously, a bug was introduced in which inherited attributes were set to the _current_ topic's attributes—which may be inherited, but may not be. It's not entirely clear _why_ that change was made, but it causes confusion because the inherited value is used as the placeholder on attribute fields. As such, if you remove an existing value from a field, that value appears to reappear as theplaceholder text.

We can fix this by simply going back to using the `DerivedTopic` version, which is how we handled this in the previous version of the editor. We'll need to be on the lookout for off-target impacts in case there was a good reason to introduce this, though. If so, we'll need to find a workaround, and carefully comment it for future versions of me.