Remove unnecessary and redundant `topic.Parent` setting
In the `Move()` action of the `EditorController`, the `topic.Parent` was explicitly being set to the new `target` parent. Not only is this unnecessary and redundant—since the underlying `ITopicRepository.Move()` will handle that—but it also introduces room for bugs since it will appear to `ITopicRepository` that the parent has not changed. I.e., if it compares `topic.Parent` to `target`, they will be the same, thus making it appear as a reordering, not a move. By not setting this in `EditorController`, we give `ITopicRepository` more reliable information, and centralize how this value is updated.