Removed explicit `id` from `SelectableTreeView`
The `SelectableTreeView`'s `OnTopicSelectableTreeView.defaults`, which are shared between multiple instances of `SelectableTreeView`, contains a default `id`. If this isn't overwritten by the client, then multiple instance of the `SelectableTreeView` will each end up with the same HTML `id`, which is invalid and error prone.

Since the `id` isn't needed for anything we're doing, and ExtJS doesn't require it, I'm just removing it. Clients can still implement it via `options` if they need to, but removing it avoids potential errors with multiple instances.