Moved the `root` definition to the `SelectableTreeView` instance
Previously, the `root` definition, which creates a new `Ext.tree.AsyncTreeNode()` instance, was defined as part of the `OnTopic.SelectableTreeView.defaults` object, which is shared between multiple instance of the `SelectableTreeView`. This effectively meant that if multiple `SelectableTreeView` instances were created on one page, they would use the same `root` node.

In practice, this meant that only the data from the first tree would be loaded into the `root` container—but that would only be _rendered_ on the last instance. That's a bit of an unexpected outcome, but regardless we shouldn't be reusing or sharing `AsyncTreeNode`s between multiple instances.

To mitigate that, I've simply moved this from the `defaults` definition and into the `SelectableTreeView` constructor.

The downside of this approach is that it doesn't allow the `root` node to be explicitly defined by the `options`, as it will get overwritten. A more sophisticated approach might be to merge it using `Ext.apply`. But as we're already merging in other defaults, and we don't expect `root` to need to be redefined for the `SelectableTreeView`, this isn't a priority. We can revisit that should it become one.

(This isn't hard to do, it's just a bit messy as it requires merging `OnTopic.SelectableTreeView.defaults` into a local `defaults`, and then merging that into the local object.)