Introduced `IsExpanded` property
The `IsExpanded` property serializes to the `expanded` property in JSON, which ExtJS utilizes to determine if the current branch should be expanded when binding to an `Ext.TreePanel` control.

I'm not loving this previous convention of exposing all properties as readonly, and then mapping them via constructor arguments. I get the reasoning for that, but we might want to revisit that at a future date. For now, I'm keeping this as is, and maintaining the convention for this new property.