Set the new `IsSelected` property as part of `MapQueryResult()`
The new `IsSelected` property allows us to construct ExtJS whether or not a branch should be expanded when binding a serialized `QueryResultTopicViewModel` graph to an `Ext.TreePanel`.

The `MapQueryResult()` function sets this value based on `MarkRelated`. If `MarkRelated` is set to `true`, and a `RelatedTopicId` and `RelatedNamespace` is defined, then the service will determine if the current node starts with the same `GetUniqueKey()` as any of the `related` topics and, if so, mark that node as `expanded`.