Introduced `ExpandRelated` topic query option
The new `TopicQueryOption.ExpandRelated` option allows callers to control whether or not an ExtJs TreePanel bound to a serialized `QueryResultTopicViewModel` should be expanded such that any nodes leading to marked related topics are expanded.

By default, this is enabled if `MarkRelated` is enabled, which in turn is automatically enabled if a `RelatedTopicId` is defined and a `RelatedNamespace` is declared. If `MarkRelated` or `ExpandRelated` are disabled, then it won't be used.