Explicitly set `Delete()` to be `isRecursive`
The `ITopicRepository.Delete()` method accepts an optional overload, `isRecursive`, which was originally set to a default value of `false`. Due to a bug in the implementation, however, this wasn't actually enforced—oops.

While this is now fixed, to maintain backward compatibility, it has been updated to a default of `true`. In the next version of the OnTopic library, however, it will be reset back to `false` as a breaking change. To prevent the need to update this at that time, it is now being explicitly set to `true` in the code.