Defer inline scripts
In preparation to defer JavaScript includes, all embedded JavaScript is being deferred by wrapping the calls in a `DOMContentLoaded` event listener. This is necessary since a) most inline scripts will rely on an external script (e.g., jQuery), and b) there isn't a way to `defer` an inline script.

While I was at it, I established a new section to the `_Layout.cshtml` called `InlineScript`—though this is of limited use since most inline script is within view components, which cannot override the master layout's sections. Still, if used, it will automatically wrap any script in that section in a `DOMContentLoaded` event listener.

For the most part, this was a straight forward exchange. In a couple of cases, concessions needed to be made to ensure that e.g. variables were available to other script blocks (e.g., by placing the variable declaration outside of the event listener block).