Defer referenced scripts
With the inline scripts deferred (09f3744) we can now safely defer the referenced scripts. This speeds up the initial page rendering time by not waiting for the JavaScript to load before rendering the DOM.

Note that we're _not_ deferring CKEditor as it's such a dominant aspect of forms that use the HTML control that it creates an unwelcome flash of unstyled content (FOUC).