Removed handling of `id` conflicts in CKEditor code
It's not entirely clear why this code is needed, as current versions of CKEditor, at least, don't mind being bound to a duplicate `id` that's within an iFrame—which is what this script was setup to avoid. Further, when placed within a `DOMContentLoaded` event listener, this throws errors. Those are probably not too hard to fix, but as this doesn't seem to serve a purpose anymore, I'm just going to remove it entirely.

As part of this, we can significantly simplify the initialization of CKEditor, as we no longer need a `setEditorInstance()` wrapper for it. Further, while we _could_ `defer` CKEditor, this creates a flash of unstyled content (FOUC) that's irritating given how prominent editor instances might be on a given form. As such, the CKEditor script is not deferred (ecfb1b5), and thus there's no reason to wrap it in a `DOMContentLoaded` event listener.

(Had we kept the original script, it still would have needed to be wrapped in a `DOMContentLoaded` event listener since the original script relied upon jQuery, which _is_ deferred.)