Added `form-control` class to backing field
Once CKEditor initializes, the `<textarea />` backing field is marked as hidden, and as such its style doesn't matter much. That said, in the brief time prior to its initialization, it will be displayed in accordance to whatever classes are applied to it. Adding `form-control` ensures that it's set to the same width as the final CKEditor window, thus reducing the impact of the flash of unstyled content (FOUC).

It's worth noting that we do _not_ want to add this to `CssClass` as that's _also_ applied to the _wrapper_, which would change the style of the final CKEditor window, and cause it to overflow over subsequent attributes. This is _exclusively_ needed for the temporarily visible `<textarea />`.

Since we're not deferring the loading of CKEditor (1052cf8) this isn't a big issue, as it renders quickly. Still, it's a simple update to help reduce any visual annoyance.