Added missing attributes for Bootstrap navs control
The navs control handles the tab functionality. There were a couple of classes and attributes which were missing, as compared to the Bootstrap documentation. These weren't strictly necessary, but I wanted to put them in since we had no reason _not_ to include them.

As part of that, I also moved the definition of `backdrop` and `keyboard` to inline `data-` attributes since they are constant settings, and aren't conditional per load. Previously, these were passed as options as part of the `modal()` initialization within `NestedTopicList.js`.