Simplified `closeModal` definition
The `closeModal` function had legacy parameters `action` and `namespace` which are no longer being used. The `action` hasn't been used in quite a while. The `namespace` was previously needed to determine the specific model window. This was replaced with the `[id^=""]` selector in a previous update, thus eliminating that need.

In addition, there's no need to call `modal('hide')` from _both_ the jQuery listener _and_ the `closeModal` function, so this is fixed. That was adding a bit of extra overhead which noticeably delayed the closing of the modal.