Ensure errors are reported for `textarea` elements
Previously, I implemented code which updates the tabs with a red asterisk if any fields within that tab have an error associated with them. My selector, however, only covered `input` fields. This is a problem if there's an error on a `textarea`, as then it's not clear why the form isn't being submitted since no tabs show up as having errors associated with them. This is easily fixed by adding the `textarea.error` class to the jQuery selector.

This was discovered as part of reducing the max length of the `MetaDescription` attribute to adhere to Google's guidance. Since a client had already entered _Description (Meta)_ tags greater than the new 160 character `maxlength`, those pages could not be submitted—but it wasn't immediately obvious _why_. This remedies that.