Ensure errors are reported for `select` elements
Previously, I implemented code which updates the tabs with a red asterisk if any fields within that tab have an error associated with them. My selector, however, only covered `input` fields. A previous commit expanded this for `<textarea />` elements (0221298); this commit expands this to also include `<select />` elements.

While we haven't had any reports of this issue, it's likely it's impacted customers, but simply not been reported. This ensures broad coverage of expected input types.