Align styling of Token Input with `.form-control`
The Token Input control's styling is different than the `.form-control`. Unfortunately, the component doesn't allow us to add our own styles. But since we have now integrated Bootstrap into our SASS file (92822dc), we can instead `@extend` the `.form-control` to apply to `token-input-list-facebook`.

For reasons I'm not entirely clear on, the styles from `token-input-styles-facebook.css`   styles—which are loaded as part of `Vendor.css`—are overwriting the styles in `Styles.css`, despite the latter being loaded second. As a result, I needed to add `!important` to a handful of attributes that are explicitly defined on `token-input-style-facebook`.

This also allowed me to fine-tune the `padding` to approximate the same height and comparable padding as used with a normal textbox. (The values are different as the text in the Token Input is wrapped in a "token", and so the padding isn't from the text, but from that token container.)