Correct XML Doc references for `Invoke()`
Originally, the `Invoke()` method was centralized as part of the `DefaultAttributeTypeViewComponent`. This became impractical since limitations of view components prevent the `Invoke()` method from being overridden. Given that, the `Invoke()` methods were distributed to each individual view component. As part of this process, however, the XML Docs were not correctly updated, and still referred back to the original `DefaultAttributeTypeViewComponent`. Oops.