Mark `DefaultAttributeTypeViewComponent` as deprecated
Originally, the `DefaultAttributeTypeViewComponent` was setup as a base class for the rest of the attribute type view components. This ended up not providing much value since limitations of view components prevent the `Invoke()` method from being overridden. Given that, the `DefaultAttributeTypeViewComponent` is no longer used in any production view components, and will be removed in the next major version of OnTopic.