Mark `DisplayOptionsViewComponent` as deprecated
Originally, the `DisplayOptionsViewComponent` provided a lookup list for targing pages to desktop or mobile as part of an early effort at supporting adaptive design scenarios. That functionality was never fully implemented, and is not actively maintained in OnTopic 4.0.0.

Indeed, while we have a `DisplayOptionsViewComponent`, there isn't a corresponding `DisplayOptionsAttribute` in `OnTopic.Metadata.Attributes`, and thus this wouldn't actually work without additional wiring to support attributes of this type being deserialized into an `AttributeDescriptor`. Further, the reference implementation doesn't define a `DisplayOptionsAttribute` content type.

Given this, we could likely remove this today without hurting anything. Technically, though, it's a breaking change if anyone is using it, so we're instead marking it as deprecated. It will be fully removed in the next major release of OnTopic.