Allow `TextAreaViewComponent` to honor `Rows` attribute
Previously, the `min-height` for `textarea` fields was set to `128px`. This overrode a `TextAreaViewComponent` with its `Rows` attribute set to e.g. `1` or `2`. That also introduced confusion when there was a `MaxLength` attribute of e.g. `160` as that meant only half of the box could be filled. By removing the `min-height` attribute in CSS, we allow the `Rows` attribute to exclusively determine the height. This can be set to `2` or even `1` if needed—though, of course, if it's `1` then the `TextViewComponent` _might_ be a better fit.