Introduced bullet list styling
The bullet list styling was being defined by ExtJS, as its CSS defaults are fairly aggressive—even though we only use it for one component. This additional CSS allows us to establish list styles for within standard attributes, while displaying them similar to ExtJS.

Note: In order to avoid competing with inline ExtJS lists, this uses a really specific selector. This won't work as is for lists that aren't directly within the `fields` class—i.e., at the top level of the attribute view. If we have a need to expand that in the future, we'll need to revisit this selector—or consider adding a specific class for this style list.