Deprecated the `CssClass` attribute
We rarely use the `CssClass` attribute, and even if we did, it's hard to imagine it being useful since there's no easy or obvious way to inject new CSS class definitions or to extend the style sheet.

If we did have additional classes that might be relevant here, they would likely relate to specific styles associated with specific attributes, and would be better suited as a `TopicList` of predefined styles defined in a `Configuration:Metadata` lookup list.