Moved `topic-info` to be sticky
Previously, the `topic-info` would end up being awkwardly orphaned in the middle of the page when there were warnings at the top, as it was marked as `position: fixed`, but the warnings were not. As such, they would scroll away, leaving an unintuitive padding at the top. By changing this to be `sticky`, it will scroll up until to the top of the page (`136px`, along with a `15px` margin). This looks much nicer.

As part of this, I was also able to get rid of the `max-width` by instead just using `width: 100%`. This wasn't previously possible with `position: fixed`.