Hide navigation on very small screen
The navigation is necessary. But, as is, the site is effectively unusable on very small screens. This update at least allows the editor to be used—even if users can't (easily) navigate to new pages without typing in the URL manually. In the future, we can revisit this by adding in a pull-out menu or something like that. Until then, this is a quick fix to keep the site at least basically usable.