Reduced the `z-index` of the `topic-info`
When applying the `sticky-top` extension, it inherited a `z-index` of 1020. This meant the topic information box shows up over the version history dropdown box. Oops.

Oddly, however, simply updating the version history to use a higher `z-index` doesn't solve the problem. In fact, the only way I could resolve it was by _lowering_ the z-index of the `topic-info` class.

It's not entirely clear why `sticky-top` comes with such a high `z-index`, nor why it can't be superceded by a higher `z-index`. But, ultimately, it doesn't really matter as we don't need or want the `z-index` on the topic info box to be high, so moving this down to a baseline value of `10` is just fine for now.