Applied C# 9.0 pattern matching for constant equality checks
C# 9.0 allows us to use `is` and `is not` in place of `==`, `!=` and `.Equals()` when comparing against constants. This might be slightly more performant, though it's primarily a stylistic preference. For `.Equal()`, this was only implemented when the comparison didn't need to be case insensitive, even if the comparison was otherwise constant, as `is` and `is not` perform an exact comparison.