Made collection properties read only
There's generally no need for a setter on a collection property—and, in fact, doing so can often lead to confusion, if not bugs. That's because the collection can still be written to without a setter, but having a setter allows the entire collection to be replaced with a reference to a different instance. To mitigate this, I've removed the getters on all collection properties.

This resolves CA2227.