Fall back to content types of parents
The `ContentTypeListViewComponent` displays a list of content types available to be created under the current topic. It bases this list on the current topic's content type. This is confusing when creating a new topic, however, since the current topic hasn't yet been saved. If the user were to quit creating the current topic, and instead choose a new topic of a different content type, it would be operating against the parent context. As such, if the `actualTopic` can't be found—because it's a new topic!—it should fall back to the parent topic instead.

This additionally fixes a bug when creating a `Container`, since in this scenario it adopts the `PermittedContentTypes` from the current topic, instead of from the `ContentTypeDescriptor`, thus allowing the same `Container` class to target collections of different content types. This previously resulted in a null reference exception, however, since `actualTopic` was `null` in this scenario.