Ensure the version history is still visible in modal windows
Previously, I added basic responsiveness to only show the version history on large screens. This inadvertantly hid the version history in the modal window for nested topics. Given this, I've adjusted what screen sizes it is visible to, so it now shows up in both large and medium sized screens. It will continue to be hidden on small screens.