Ensure the version history can fit the widest date format
The version history dropbox is generally the right size. There are some cases, however, where the data wraps to a new line, which makes the versions harder to read, and looks unprofessional. E.g., `2020-12-31 12:00:00 AM`. This is especially an issue because a recent update to the database allows compressing all versions between a range; in that case, they will default to `12:00:00 AM`—and, if done for the entire year, `12-31` of that year. That almost always forces the wrapping.

This fixes that issue, while also cleaning up the styles a bit, as there were some unnecessary styles declared inline in the `_Toolbar.cshtml` template.