Use `Collection<>` instead of `List<>`
Publicly visible members should return `Collection<>` instead of `List<>` as it offers a more comprehensive and familiar API, whereas `List<>` is a lightweight interface built for performance over usability.

This is a bit of a false positive in that these classes are only used internally and, therefore, don't strictly need to be public; the reason they are is an accident of history, due to the models being stored in a separate assembly. That will be fixed in the next major version of the OnTopic Editor. In the meanwhile, this knowledge gives us confidence that we can change the type without impacting the implementation.

It's rare that changing the type from `List<>` to `Collection<>` will break callers, since the latter implements _most_ of the same members as the former, and both implement `IList<>` used by the `TopicMappingService`. Still, it is possible if callers explicitly define the type, or if they call a handful of members which are specific to `List<>`.

The most common of these in our code is `AddRange()`, a fast method for adding a `List<>` to a `List<>`. In those cases, the code in `OnTopic.Editor.AspNetCore` has been refactored to instead use a `foreach()`.

This mitigates CA1002.