Throw `InvalidOperationException` instead of `Exception`
Classes should not throw `Exception`, but rather more specific exception types. This is only done in one case, as part of the `StandardEditorComposer`'s `ActivateEditorComponent()` method. As `InvalidOperationException` is more specific than `Exception`, this shouldn't cause any backward compatibility issues. And, regardless, as this exception is intended to alert developers about an invalid configuration, it's highly unexpected that developers would be deliberately capturing this exception in their code. This resolves CA2201.