Removed redundant condition
It's likely that this was _intended_ to validate `AttributeValue` instead of `AttributeName`. Introducing that now would be a potentially breaking changes for cases where a caller is filtering by topics that do _not_ define an attribute value. That seems like a potentially valuable capability regardless. Given that, I'm removing the redundant condition, but not "correcting" it to the presumed intent. This resolves an instance of CA1508 (though there remain other instances that aren't as obvious or necessary to fix).