Installed NuGet version of NetAnalyzers
With the migration from `FxCopAnalyzers` to `NetAnalyzers`, we no longer _need_ to use the NuGet package, as the analyzers are now built into the SDK. Unfortunately, that means that they're often lagging behind the latest fixes. When we originally tried installing the NuGet package, we got duplicate warnings, and so had relied on the SDK implementation. That appears to have been subsequently resolved. As such, I'm moving back to the NuGet implementation to ensure we're always using the latest version.