Updated to alpha release of OnTopic 5.0.0
Note: The alpha release is mislabeled as 4.7.0; in practice it will be released as 5.0.0, due to an abundance of breaking changes, which has been planned for this release.

This is a prelease, and we expect there to be bugs. Indeed, the goal of this branch is to establish basic integration testing of the library to help begin the update of OnTopic Editor, while still ironing out any outstanding issues with the alpha release.