Removed chaining of `TopicViewModelLookupService`
Previously, extensions of any of the default `ITypeLookupService`s—such as the `TopicViewModelLookupService`—had to be chained to ensure that the final implementation had access to all types from all libraries.

This meant the implementors of `OnTopic.Editor.AspNetCore` _had_ to maintain a dependency on `OnTopic.ViewModels`, even if the actual implementation didn't actually require it, as otherwise there wasn't an (easy) way to incorporate both the editor's view models and the reference view models into a single lookup service.

This challenge is mitigated by the introduction of the `CompositeTypeLookupService`, which allows multiple `ITypeLookupService` implementations to be added to a single implementation, without needing to be chained.

As a result, we can remove the chaining, and instead derive directly from `StaticTypeLookupService`.

This is potentially a breaking change; it will require that implementors that also depend on the `TopicViewModelLookupService` wrap it and the `EditorViewModelLookupService` into a new `CompositeTypeLookupService`. In doing so, however, it will make it easier for implementors to combine disparate lookup services into a single implementation without needing to chain them together. It will also allow implementors to remove their dependency on `OnTopic.ViewModels` if they don't otherwise need it.