Migrated `ReadOnlyTopicCollection<T>` to new `ReadOnlyTopicCollection`
The original `ReadOnlyTopicCollection<T>` has been renamed to `ReadOnlyMappedTopicCollection<T>`. But we don't actually want that when it comes to `Topic.Relationships`, because it artificially restricts the topics in the collection from having the same `Key`.

Given that, the `Topic.Relationships` collection now returns mapped `ReadOnlyTopicCollection` instances instead of `ReadOnlyMappedTopicCollection<T>` instances. That allows relationships to point to multiple unique `Topic` entities with the same `Topic.Key`.

As such, this not only addresses the change in the base class name, but simultaneously addresses the change in the `Relationships` data model—and permits `Topic.Relationships` to have references to any topics, not just topics with unique `Key`s.