Updated to alpha release of OnTopic 5.0.0 (cont.)
In a previous commit, I updated to the alpha release of OnTopic 5.0.0 (967152e). That inadvertantly only included the `OnTopic` library, but not ancillary libraries such as `OnTopic.Data.Sql` or `OnTopic.ViewModels`. Whoops! This remedies this.

Note: The final release will be 5.0.0, but the current preview release is (incorrectly) tagged as 4.7.0.