Handle content type during search
As part of OnTopic 5.0.0, core attributes—such as `Key`, `ContentType`, and `ParentID`—are all moved from `Topic.Attributes` to `Topic`. As such, the common requirement of filtering topics by `ContentType` using an `AttributeKey`/`AttributeValue` pair won't work.

To mitigate that, we intercept the request in the `TopicQueryService`, and derive the value for content type from the `Topic` instead of from `Topic.Attributes`.