Handle key when constructing the `EditingTopicViewModel`
As part of OnTopic 5.0.0, core attributes—such as `Key`, `ContentType`, and `ParentID`—are all moved from `Topic.Attributes` to `Topic`. As such, the `Key` attribute's value can no longer be sourced from the `Topic.Attributes` collection.

To mitigate that, a condition was added to detect the `Key` attribute, and pull the value from `Topic.Key` instead.