Introduce support for Topic References
In OnTopic 5.0.0, we're introducing the concept of topic references—which are like relationships, but have a 1:1 relationship per key, instead of a 1:n relationship per key.

We already have the concept of `ModelType.Reference` accounted for in the data model, but there wasn't special handling of it in `EditorController`. This update ensures that the value of `ModelType.Reference` is pulled from `Topic.References.GetTopic()`, and set via `Topic.References.SetTopic()`—with the latter using a new `SetReference()` helper function.